refactor(frontend): extract useTabRouting composable (P2-2 F8)
Move the tab<->URL route sync functions (syncRouteWithTab,
clearRouteTabQuery, clearCreateQuery, syncFromRoute) out of DataStudioNew.vue
into composables/useTabRouting.js. The composable resolves its own
useRoute/useRouter and injects shared tab/catalog state, lazy-wrapping the
forward refs (loadSchemas/loadTables/openTableTab). Tab-lifecycle handlers
that border on the catalog/open-tab core (F7) are intentionally left in the
component for now.

Verified: mount smoke green, npm run build, npm run lint (0 errors); orphan
check clean; route/router still used by remaining component code. Functions
are verbatim-moved.
Refs docs/plans/2026-06-18-frontend-megacomponent-split-plan.md (F8).
